requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
olefile==0.44
Pillow==4.3.0
requests==2.18.4
selenium==3.7.0
urllib3==1.22
